First Long Time Fuel Trim Implementation
For this the ECU must be always on 5V

It's based on STFT and an bias curve. The correction is linear if STFT greater than as threshold defined by: (1-10^ (-20*abs(1-STFT))), so, the correction must be stable;

We have some limits on the correction, using CLT sensor, of using the LTFT table and modifying it.

Was implemented a delay on  the correction after fast acceleration by TPS, as it could be an problem to the correction.

Some minor details...